Move directory table into right file
